+++ PokemonTests/LoadPokemonFromRemoteUseCaseTests.swift
@@ -33,26 +33,6 @@ public final class RemotePokemonLoader {
     }
 }
 
-class HTTPClientSpy: HTTPClient {
-    private var messages = [(url: URL, completion: (HTTPClient.Result) -> Void)]()
-
-    var requestedURLs: [URL] {
-        return messages.map { $0.url }
-    }
-    
-    func get(from url: URL, completion: @escaping (Result<(Data, HTTPURLResponse), Error>) -> Void) {
-        messages.append((url, completion))
-    }
-    
-    func complete(with error: Error, at index: Int = 0, file: StaticString = #file, line: UInt = #line) {
-        guard messages.count > index else {
-            return XCTFail("Can't complete request never made", file: file, line: line)
-        }
-        
-        messages[index].completion(.failure(error))
-    }
-}
-
 final class LoadPokemonFromRemoteUseCaseTests: XCTestCase {
 
     func test_init_doesNotRequestsDataFromURL() {
